Split containers into sub-containers
