+++ app/javascript/react/components/Loading.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Loading = props => {
+  return (
+    <h4>Please wait, system processing...</h4>
+  )
+}
+
+export default Loading
