+++ app/javascript/react/containers/ToyNewContainer.js
@@ -2,6 +2,7 @@ import React, { useState } from 'react'
 import { Redirect } from 'react-router-dom'
 
 import ToyForm from "../components/ToyForm"
+import Loading from "../components/Loading"
 
 const ToyNewContainer = props => {
   const [newToy, setNewToy] = useState({})
@@ -57,7 +58,7 @@ const ToyNewContainer = props => {
     <ToyForm
       submitForm={submitForm}
       defaultFormData={defaultFormData}
-      />
+    />
   )
 }
