+++ app/javascript/react/containers/ToyShowContainer.js
@@ -1,33 +1,14 @@
 import React, { useState, useEffect } from "react"
-import { Link } from 'react-router-dom'
 
 import ShowTop from "../components/ShowTop"
-import UpForGrabs from "../components/UpForGrabs"
 import ToyShowDetails from "../components/ToyShowDetails"
-import AddToToyboxButton from "../ui/AddToToyboxButton"
+import Loading from '../components/Loading'
+import UpForGrabsContainer from './UpForGrabsContainer'
+import ToyButtonsContainer from './ToyButtonsContainer'
 
 const ToyShowContainer = props => {
-  const [toyData, setToyData] = useState({
-    id: null,
-    toy_name: "",
-    manufacturer_name: "",
-    min_age: null,
-    max_age: null,
-    toy_photo: { hero: { url: null } },
-    upc: null,
-    description: "",
-    toyboxes: [
-      {
-        user: {
-          id: null
-        }
-      }
-    ],
-    current_user: {
-      id: null
-    }
-  })
-  const [toyAdded, setToyAdded] = useState(false)
+  const [toyData, setToyData] = useState({})
+  const [loading, setLoading] = useState(true)
 
   const getToyInfo = () => {
     fetch(`/api/v1/toys/${props.match.params.id}`)
@@ -42,29 +23,8 @@ const ToyShowContainer = props => {
     })
     .then(response => response.json())
     .then(parsedData => {
-      setToyData(parsedData.toy)})
-    .catch(error => console.error(`Error in fetch: ${error.message}`))
-  }
-
-  const addToToybox = event => {
-    event.preventDefault()
-    fetch(`/api/v1/toys/${props.match.params.id}/toyboxes`, {
-      credentials: "same-origin",
-      method: "POST",
-      headers: {
-        Accept: "application/json",
-        "Content-Type": "application/json"
-      }
-    })
-    .then(response => {
-      if(response.ok) {
-        setToyAdded(true)
-        return response
-      } else {
-        let errorMessage = `${response.status} (${response.statusText})`
-        let error = new Error(errorMessage)
-        throw(error)
-      }
+      setToyData(parsedData.toy)
+      setLoading(false)
     })
     .catch(error => console.error(`Error in fetch: ${error.message}`))
   }
@@ -73,56 +33,38 @@ const ToyShowContainer = props => {
     getToyInfo()
   }, [])
 
-  const availableList = toyData.toyboxes.map(toybox => {
-    if(toybox.for_sale) {
-      return(
-        <UpForGrabs
-          key={toybox.toy.id}
-          toybox={toybox}
-          getToyInfo={getToyInfo}
-          currentUser={toyData.current_user}
-        />
-      )
-    }
-  })
-
-  const details = <ToyShowDetails
-                    description={toyData.description}
-                    manufacturerName={toyData.manufacturer_name}
-                    upc={toyData.upc}
-                    minAge={toyData.min_age}
-                    maxAge={toyData.max_age}
-                  />
-
-  let toyAddedStatus = ""
-  if(toyAdded) {
-    toyAddedStatus = <p>Toy added to your toy box!</p>
-  }
-
-  let toyInToybox = true
-  if (toyData.current_user && toyData.toyboxes.filter(toybox => toybox.user.id === toyData.current_user.id).length === 0){
-    toyInToybox = false
-  }
-
-  return (
-    <div className="grid-y grid-margin-y">
-      <div className="cell">
-        <ShowTop
-          name={toyData.toy_name}
-          photo={toyData.toy_photo.hero.url}
-          details={details}
+  if(loading){
+    return <Loading />
+  } else {
+    return (
+      <div className="grid-y grid-margin-y">
+        <div className="cell">
+          <ShowTop
+            name={toyData.toy_name}
+            photo={toyData.toy_photo.hero.url}
+          >
+            <ToyShowDetails
+              description={toyData.description}
+              manufacturerName={toyData.manufacturer_name}
+              upc={toyData.upc}
+              minAge={toyData.min_age}
+              maxAge={toyData.max_age}
+            />
+          </ShowTop>
+        </div>
+        <div className="cell text-center">
+          <ToyButtonsContainer toyData={toyData} match={props.match}/>
+        </div>
+        <div className="cell">
+          <UpForGrabsContainer
+            toyboxes={toyData.toyboxes}
+            getToyInfo={getToyInfo}
+            currentUser={toyData.current_user}
           />
+        </div>
       </div>
-      <div className="cell text-center">
-        {!toyInToybox && <AddToToyboxButton addToToybox={addToToybox}/>} {toyData.current_user && <Link to={`/toys/${toyData.id}/edit`} className="button">Edit</Link>}
-        {toyAddedStatus}
-      </div>
-      <div className="cell">
-        <h2>Up for Grabs</h2>
-        {availableList}
-      </div>
-    </div>
-  )
+      )
+  }
 }
 
 export default ToyShowContainer
